Fix for PDI-687 : Repository Explorer deletes directory without warning if there are only subdirectories, no jobs or transformations
git-svn-id: svn://source.pentaho.org/svnkettleroot/Kettle/branches/3.0.2@6481 5fb7f6ec-07c1-534a-b4ca-9155e429e800

+++ src/org/pentaho/di/repository/RepositoryDirectory.java
@@ -509,7 +509,8 @@ public class RepositoryDirectory
 		{
 			String trans[]   = rep.getTransformationNames(getID());
 			String jobs[]    = rep.getJobNames(getID());
-			if (trans.length==0 && jobs.length==0)
+			long[] subDirectories = rep.getSubDirectoryIDs(getID());
+			if (trans.length==0 && jobs.length==0 && subDirectories.length==0)
 			{
 				rep.deleteDirectory(getID());
 			}
